Update README.md
The introductory R course at the University of Bern's Department of Biomedical Research (DBMR) provides students with a foundational understanding of R programming and its applications in the field of biomedical research. This course offers an immersive learning experience, equipping participants with the essential skills needed to manipulate data, perform statistical analysis, and generate visualizations using the R programming language. Through hands-on exercises and practical examples tailored to the specific needs of biomedical research, students gain proficiency in data analysis techniques, enabling them to make informed decisions and contribute effectively to scientific research endeavors. The DBMR's introductory R course serves as a valuable resource for students aspiring to excel in the intersection of programming and biomedical sciences.